+++ app/rdev/http/InvalidCSRFTokenException.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Copyright (C) 2015 David Young
+ *
+ * Defines the exception thrown when the CSRF token is invalid
+ */
+namespace RDev\HTTP;
+use Exception;
+
+class InvalidCSRFTokenException extends Exception
+{
+    // Don't do anything
+}